完善单元测试 添加消息转换器

+++ ping-business/business-user-web/src/test/java/cn/procsl/business/user/web/controller/AcceptedTestController.java
@@ -31,7 +31,7 @@ public class AcceptedTestController {
         return 200D;
     }
 
-    @PostMapping("double")
+    @PostMapping("long")
     @Accepted
     public long postLong() {
         return 3000L;
